[inflight] Use a clonable Sender rather than an Arc<Mutex>
The only purpose of doing the Arc<Mutex> dance, or the even more 
complicated ArcSwapOption dance, as I briefly went with writing this 
commit, was to make Sender Send+Sync+Clone.

A better solution is to use a channel that is already designed to have 
these guarantees built-in. As a bonus, crossbeam's implementation is 
reportedly more performant.